Move special controls.
